feature: ANSI escape sequences support
Woohoo! Most ANSI escape sequences now work! Things like underline or
bold don't, but move cursor, clear screen, change foreground and
background color do! The function doing that is hooked directly in the
tty driver in the term_putchar() function. In the future we shall make
this a lib or something so it can be reused in another terminal.