feature: add full paging support in higher half
I implemented (with the great help of osdev.org) the recursive page
directory trick to manage page tables when the kernel is in higher
half, so now any virtual address can be mapped to any physical address
with a simple call to paging_map_page().
I also reduced the memory usage of the kernel by removing some static
stuff.

Onwards to greater things!