basic usermode code execution
The main thing here is the introduction of proc_run_code which sets up a
few pages of memory for the supplied code and its stack in a new page
directory and then jumps to this code in usermode. For now, the code
that runs just sets `eax` to 42 and loops.
I introduced preliminary structures to manage processes, mainly
process_t.
I needed to allocate 4 MiB of physical memory aligned on a 4 MiB
boundary, and in doing so I found and fixed a bug in my physical memory
manager.