basic cooperative multitasking
I'm pretty sure there's a bug currently with `eip` not being saved
between context switches, but contexts _do_ switch so that's worth a
commit.
My previous algorithm for insert processes in the list was ridiculous,
in O(n) when constant time is trivial to get.
My IDT entry for interrupt 48 wasn't setup correctly, it didn't have the
proper flags to be triggered from a userspace process.