use separate kernel stacks for each process
The previous solution was complete garbage, I had put no thought into it
at all. Now everytime execution switches to kernel mode, the stack is
set to a fixed location, that of the process' kernel stack.
Now what happens when an interrupt happens (e.g, from the PIT) while the
execution was already in kernel mode (e.g from a syscall) ? I'll have to
check but hopefully the kernel stack isn't loaded from the TSS then,
it's kept as-is.