putchar syscall, dynamic PIDs
Four years after its inception, I'm proud to announce the first
>Hello, World
from one of SnowflakeOS' userspace process. This is the Hello world that
counts. After all, a kernel is there to be used, and what pride is there
in printing text for oneself?
As a side note, my `wait` syscall implementation has regressed, I need
to think a bit more about schedulers to do it right.