allocate code and stack pages properly
We no longer use large pages for the code and stack pages of user
processes. Instead, a process is given 40 KiB of stack and just enough
pages for its instructions.
This led to a design change in `paging_get_page`, as page tables need to
be allocated with user flags if necessary, and modifying that after the
fact can get pretty ugly, if for instance the code (or stack) covers
several page tables.