document and improve GDT, IDT and ISR code
`idt.c` is now solely responsible for setting up the pointer to our IDT,
and other modules such as `isr.c` and `irq.c` are responsible for
filling it with handlers. This makes reading `isr.c` a lot less
confusing.
Fixed potential issues with some ISR handlers declared as `ISR_NOERR`
when they did in fact push an error code. That issue came from James
Molloy's tutorial a long time ago, and was spotted thanks to the osdev
wiki page about it. In fact, it's that page that prompted me to go over
all that code, and it's made a lot of things clearer, such as the state
of the stack, segment registers and `eflags` when an interrupt is
called.
There's still a major problem with my context-switching code, as it
doesn't restore working registers. I had a poor understanding of how my
interrupt handlers were called, how I could return from them and what I
would gain in doing so, which lead to me trying to hack a solution.