renamed tty_* to term_*, few cleanups
`printf` should now be accessible from my sort-of-user-libc, of which I
don't remember much. I'll have to dig into the buildsystem, but my next
step is to be able to compile C programs and load them as GRUB modules
to run them in userspace.
Also, SnowflakeOS finally builds warning-free, thanks to a small comment
that disables fallthrough warnings where appropriate.