revamp build system
It's now make-based, and imo much clearer. It's just `make` to build
everything and `make qemu/bochs` to run.
In a first pass, headers are copied to the fake root directory, and on a
second pass the different parts of the OS are compiled, and have access
to all previous headers.
I've added a `modules` part containing code to be loaded as a GRUB
module. I haven't been able to get `ld` to honor the flat binary output
format, so I guess I'll have to write an ELF loader pretty soon if I
ever want to run C code in userspace.