quick fix to proc_enter_usermode
How, you ask, did I not notice this obvious bug before?
Well, the reason is that by extraordinary chance, it just worked
without. The function epilogue popped enough values off the stack for
%esp to end up pointing at 0x0 before `ret`, so the execution still
jumped to the code, just not in usermode. On subsequent switches, the
situation was back to normal. Cool bug.