PS/2 controller and mouse driver
The PS/2 controller (or controllers) is now properly initialized, and so
are connected devices, if any.
The goal was mouse support, and we now have it! The sensitivity on the
y-axis seems far greater than on the x-axis though.
The keyboard handling still has to be vastly improved.
Fixes an age-old bug with an unhandled IRQ0 during the boot sequence.