near-proper keyboard driver
Scan codes are properly translated into key codes, and there is a map
from relevant key codes to printable characters. The driver is 'aware'
of key events and also keeps track of simultaneously pressed keys.
I'm really not sure how the kernel will communicate keyboard state and
events to userspace. I just had an idea: have the stream of key codes be
a file, like /dev/keyboard. In ToaruOS, scan codes are available as a
file and the translation is done in userspace. I could do that too if I
find any advantage to it.
Few random cosmetic fixes.