fix y-axis mouse movements, proper shift key
To be honest, I don't get what was wrong with how I handled vertical
mouse mouvements. It's some two's complement magic, of course. The issue
was that mouvements were reported as either '-256' or '+255' instead of
'-1' and '+1'.
For the keyboard changes, this kind of key code to character translation
will surely be moved to userspace soon enough, given that it all depends
on keyboard layout...