Framebuffer rendering from userspace
I've introduced snowlib, the system library for SnowflakeOS. It exposes
the syscalls that don't fit into libc. For now there's a rough
equivalent of `sbrk` called `snow_alloc`, and `snow_render` that allows
drawing a full userspace framebuffer to the screen.
My rough plan is to build a graphics library in snowlib and a compositor
in the kernel.