Fixed program prelude not being used
In a previous commit I noted how the content of `start.S` always got
placed after a compiler-generated call to `main` in the binaries of my
modules, even though the linker script should have put it at the
beginning.
Well, this is fixed now, and I'm not entirely sure why. There must have
been a problem with `start.o`, the linker script referenced it but that
object may not have been recompiled in a long time. Indeed, I deleted it
and the build failed, and fixing this fixed the bug.