Development environment cleanup
On one hand, VSCode started failing to resolve headers, or using
wrong/outdated headers for autocompletion, which was driving me crazy.
On the other hand, my precious buildsystem failed in a new way: a change
in `snow` didn't trigger the recompilation of `modules`. This is in fact
normal, and I've fixed it by forcing the recompilation of modules...