Fonts, malloc, a pinch of GUI
Documentation is very much lacking concerning the PSF1 font format. Good
thing it's the simplest format in existence.
There's a `malloc` but there's no `free`. A common theme in SnowflakeOS.
Somehow, and I must insist, _somehow_, the framebuffer now works in ARGB
instead of RGBA, and there no longer is a random line of white pixels at
the top of the screen. Either I've fixed something memory-related by
inadvertance, or qemu self-updated in the background...