Implement a window manager
Applications can register a window object of a specified size with the
WM. The application is then expected to call `wm_render_window` in a
loop. That function will draw the window to the off-screen buffer iff
windows with lower z-order have been draw already, thus ensuring that
windows are drawn in the correct order.
As such, the framerate is limited by the framerate of the slowest
application.
The GUI shown at this point is pure mockup.