Hunting down undefined behavior
Compiling with O2 shifts my framebuffer to the right in bochs, for some
reason.
At some point, adding a printf statement in wm_open_window fixed the
symptoms...

Cause: depending on the optimisation level and the emulator used, the
background window happened to be opened after the other window, and the
window manager set its x coordinate to (lowest x + 50), thus shifting
everything to the right.

This is my first bug due to userspace interaction! I really need to fix
that window positioning code, it's terrible, and it looks hard to do
correctly.