FPU support
The FPU is now initialised and its context is saved per-process.
States are saved and restored each context switch, we're not doing smart
things here.
We may even be doing dumb things: the first time a process is switched
to, the FPU context restored is pretty much garbage. Hopefully this is
all cleared up when said process starts doing calculations. Otherwise
it'll make for fun debugging times later on.