Fix the physical memory map overwriting modules
It turns out GRUB places modules after the kernel in memory. I did not
anticipate this, and thus I was using the memory right after the kernel
to store the bitmap used in the physical memory manager.

A strange bug led me to this fix:
 When using the new memory allocator and memory mapping the first 16 MiB
 of memory, the letter 'w' in the "loop.c" program got scrambled. Just
 that one letter. And no bug whatsoever if I identity mapped only 4 MiB.

Mapping pages allocates physical memory for new pages tables, so my
guess is that that was it.