Speed up window rendering
This commit makes a huge difference, I wish I wasn't lazy enough to
measure it. Basically, initial drawing is instant in qemu, and as much
quicker in bochs.
The reason: the framebuffer was still being copied in its entirety on
single window redraws. Also, redrawing parts of windows on top of the
window we wish to redraw is useless, and no longer done.