Window dragging, glitch fixes
Windows can be dragged from anywhere on their surface; it still allows
for click events (which aren't implemented as of now).

Some time was spent investigating why dragging a window up left a trail
of one-line garbage at the bottom. It was of course an off-by-one,
caused by me not using the tools I made to avoid them, in this case the
function to get the `rect`.
There is still a glitch escaping me: when dragging a window over
another, it sometimes - and only sometimes - leaves traces behind. Same
thing with the cursor, when it's close to a window border.