Fix clipping bug in specific dragging conditions
In this scenario, dragging the second window two pixels up across the
    +----+     bottom border of the first window left a part of its
    |1 +----+  bottom row in place, i.e. not drawn over by the
    |  |2   |  background. It was difficult to reproduce, so I ended
    +--|    |  up mocking user input on subsequent timer interrupts.
       +----+
It's all my fault, I had replaced inequalities with strict inequalities
in `rect_split_by` to fix a bug, but it was the wrong fix. Then, I fixed
the root cause of the first bug, and this new one popped up. It took me
around two days to track down.