Click events passed onto windows
And the event framework is laid out! Windows can request an event
structure which contains whatever happened to the window since the last
event retrieval, for now just clicks, but there are several event types
and an event struct can contain exactly one of each.
Also, windows can now ask to redraw only a part of themselves, which
will allow for good performance for applications that need it. Paint?