Implement a terminal with a basic shell
This terminal is our `init`, so to speak. It knows a couple of builtin
commands, and what it doesn't recognize it calls `exec` on.
It's worth mentionning that it doesn't (and can't) keep track of what
programs it launched, and therefore doesn't print their standard output.
That would be so cool, I think I'll do that next.
Now, the thing is we no longer have a background running by default...