Fix static variables overlapping with memory pool
Alright, I learned something today:
 - flat binaries can use addresses past their size to store static variables,
 - there's no way to tell how much memory a flat binary expects to have
   for static variables.

The bug was very much related to the aforementionned cool facts. It so
happened that my terminal program was 0x2ff1 bytes long, juuust short of
three pages, and it placed the global, static variable `used_memory` at
address 0x3000. But when loading the program, I allocated three pages
for it, so that program's malloc memory pool ended up starting at...
0x3000 exactly. On malloc(n), `used_memory` was increased by n, thus the
first block's `next` member got assigned n instead of staying null. What
does the next allocation check? If the previous block has a successor.
Guess what? it does, it's located at... n. And so, the allocator
returned an address corresponding to garbage at the beginning of the
program's code... Ah, the marvels of osdev.