Large file support, missing Makefile
The ext2 driver can now read files larger than 12KiB, and this ability
coupled with using a disk image just the right size have made it
possible for the wallpaper to be loaded from disk.
Also, I had forgotten to push the Makefile for the new and ill-named `ui`
library.