ext2 block allocation, pmm bug fix
Slowly moving towards a functional ramfs with the ability to mark a
block as used.

The real reason for this commit is the bug fix though: while figuring
out a better alternative to my stupid `equal_n` function, I added
`strncmp` to my libc (I had forgotten the existence of memcmp...), and
suddenly, my terminal stopped working.
Armed with the experience of recent bugs, I got the feeling it might be
related to my out-of-binary static variables being non-zero at program
launch, and added that guarantee. It fixed the problem.
However, paint was still crashing at launch: garbage program code being
executed. Tracing where the corruption occured, the paging code appeared
to be the culprit, and I debugged that for too long without success. A
day later, it occured to me that my paging code being correct, it could
access an invalid virtual address. Physical pages were affected, so
perhaps they were mis-attributed? A quick 2 a.m. code review of the code
responsible for marking a span of physical memory as used led me to
notice that the last block to mark wasn't being marked unless the span's
size was a multiple of the block size... It fixed things up.