Reworked paint to use new ui library
It was surprisingly easy and not buggy!
One point of friction in the library identified: it's difficult to size
a container in its other direction according to, say, the max size of
its elements in that other direction. The main direction of the
container being vertical for a vbox, its other direction being
horizontal.