Allow grub modules of arbitrary sizes
What we used to do was this: put our page allocation bitmap after the
kernel and its modules in memory. This one of the very first parts of
the kernel, made back in 2015, and this was never given more thought.
Presumably, my logic was that this bitmap was of variable size,
depending on the RAM available, thus I needed to allocate it
dynamically. There lies my mistake: even assuming we have the maximum
amount of RAM available, 4 GiB, the bitmap still takes a reasonable 128
KiB of memory that can be statically allocated.
This dumbness gone, we can map just the right amount of memory for grub
modules, so that they can be copied in peace.

Now the question is, what will motivate me to write a disk driver?