Moving towards a VFS, part two
The `fs` module is now independent of `ext2`, and inodes can
theoretically be on different filesystems in the same tree, aka mount
functionality.
Also, fixed a bug with folders being assigned to the files list of
inodes in the tree... It worked anyway, but could cause invalid casts in
weird cases that started happening with this change.