Make the buildsystem more consistent
Used to be, when you made a change in the libc, the kernel didn't get
rebuilt by a call to `make`. Now it does! I _will_ get the hang of
Makefile good practices.
Still an inconvenience in the way modules (aka apps) are built: they
each have all of the other as dependencies, so a change in one triggers
a rebuild of all of them.