Paint: save functionality
The pathname is fixed right now, this GUI isn't there yet to show a
modal with a textfield. Better rename the saved file if you intend to
save several in one session. Oh wait, you can't rename a file yet!

Relatedly, fixes a memory leak when writing to a file on an ext2 file
system. I found it because at first I wrote the saved image pixel by
pixel... Which took several seconds, and made the kernel abort due to
being out of memory.