Working stdout implementation
A process can now declare itself to be a terminal with the `maketty`
system call. This gives it a new file descriptor corresponding to
stdout, that is then inherited by the processes it executes, and the
processes those execute, etc...
It's the terminal's job to read from stdout (a strange concept, I know)
and do something with it.
`stdout` is backed by a circular buffer; old content may be overwritten
if not read.