Fix another make issue with modules
Subprojects that weren't dependencies of other things - modules and
doomgeneric (which should be among modules but whatever) - failed to
trigger a rebuild of the ISO when modified.
This dependency is handled by the main Makefile, which sees changes
based on directory modification dates, which is itself updated through
`touch` in the modules' Makefile and doom's.