Switch to multiboot2
Yay, We're slightly more modern! This grants us the ability to access
ACPI tables easily! Baby steps towards PCI-IDE.

This uncovered a bug (another!) in the pmm: it was protecting more
memory than it should, and correcting this made things fail. The
calculations for the number of blocks to protect given a region didn't
account for regions crossing block boundaries...