wm api changes now that events are independent
In particular, `wm_event_t::type` is no longer a bit field but a proper
value, and using `type & WM_EVENT_...` to check the event type no longer
makes sense.
I couldn't bring myself to use a union for the event contents... It'd
save bytes, but it adds one more level of indirection when extracting
the content, and that's just too much.