Documentation for AHCI
I'm slowly learning all the bits and pieces of this whole PCI/AHCI/SATA
architecture.
To sum things up, at the top we have PCI, a general purpose bus. On this
bus one may find one or more AHCI controllers, which can be thought of
as hubs for disks (and more obscure devices). Each disk plugs into a
port of an AHCI controller, and the set of commands it will respond to
is more or less called SATA (ATA, ATAPI might pop up here and there
too).
Note that SATA isn't really something you use, it's just definitions.
What you use to communicate with disks is the AHCI controller, whose
role is to coordinate data transfers.
Each port (where a disk is plugged) has a memory-mapped command list and
'reply' structure (FIS received).
When a command is done, an interrupt may be fired, and it says for which
port(s) it was. The AHCI driver can then look at the reply structure for
these ports, and find some kind of ACK or NAK, and whatever buffer was
prescribed by the command list will be filled (in case of a read) by
that point. Reads and writes can just as well be made blocking, which is
the case in this driver.