wip / reset-me: here's the plan, nothing done
Have ext2 driver allocate a cache-disabled page to use with read_block
and co.
If it was needed, accesses to it could be mutex-protected.
I think there _is_ a bug somewhere, and that the current solution of
invalidating pages / disable cache on those random kmalloc'd pages from
the ext2 driver actually _should_ work. That it gives weird crahes, only
for one userspace program, is just weird.
I'm hoping the one-page-for-io paradigm makes some things cleaner and
maybe shine some light on this.